Moved data sources to one-per-source-file.
